backend/requirements.txt
sqlalchemy
passlib
fastapi
sqlalchemy
uvicorn
PyMySQL
PyJWT
python-multipart
mysqlclient
